# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /public/jobs/{token}:
    get:
      operationId: getPublicJob
      tags: [public]
      summary: Get a customer-facing job status by share token
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public job status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicJob"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /public/quotes/{token}:
    get:
      operationId: getPublicQuote
      tags: [public]
      summary: Get a customer-facing quote by share token
      description: |
        Returns the quote for the public acceptance page. The first time a Sent
        quote is fetched it is marked as Viewed.
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicQuote"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /public/quotes/{token}/accept:
    post:
      operationId: acceptPublicQuote
      tags: [public]
      summary: Accept and e-sign a quote from the public link
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteAcceptInput"
      responses:
        "200":
          description: Accepted quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicQuote"
        "400":
          description: Invalid request or quote not acceptable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /public/quotes/{token}/decline:
    post:
      operationId: declinePublicQuote
      tags: [public]
      summary: Decline a quote from the public link
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Declined quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicQuote"
        "400":
          description: Quote not in a declinable state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    JobStatus:
      type: string
      enum:
        - New Lead
        - Estimate Scheduled
        - Estimate Completed
        - Proposal Sent
        - Approved
        - Material Ordered
        - Material Received
        - Scheduled
        - In Progress
        - Final Walkthrough
        - Completed
        - Invoiced
...
    CrewAssigned:
      type: string
      enum: [Crew A, Crew B, Crew C, Crew D, Unassigned]
    FlooringType:
      type: string
      enum:
        - Carpet
        - Hardwood
        - Tile
        - Laminate
        - Luxury Vinyl Plank (LVP)
        - Luxury Vinyl Tile (LVT)
        - Waterproof Flooring
        - Commercial Carpet Tile
        - Commercial LVT
...
    ProposalStatus:
      type: string
      enum: [Draft, Sent, Viewed, Accepted, Declined]
    DepositType:
      type: string
      enum: [none, percent, amount]
    InvoiceStatus:
      type: string
      enum: [Draft, Sent, Paid, Overdue]
    InvoiceLineCategory:
      type: string
      enum: [Labor, Materials, Add-on]
...
    Room:
      type: object
      required: [id, name, length, width]
      properties:
        id:
          type: string
        name:
          type: string
        length:
          type: number
        width:
          type: number
        scope:
          type: string
    StageEvent:
      type: object
      required: [stage, at]
      properties:
        stage:
          $ref: "#/components/schemas/JobStatus"
        at:
          type: string
    JobPhoto:
      type: object
      required: [id, jobId, stage, objectPath, caption, createdAt]
      properties:
        id:
          type: string
        jobId:
          type: string
        stage:
          $ref: "#/components/schemas/JobStatus"
        objectPath:
          type: string
        caption:
          type: string
        createdAt:
          type: string
...
    PublicCompany:
      type: object
      required: [companyName, phone, email, website]
      properties:
        companyName:
          type: string
        phone:
          type: string
        email:
          type: string
        website:
          type: string
    PublicJob:
      type: object
      required:
        - jobNumber
        - customerName
        - city
        - flooringType
        - status
        - stageHistory
        - photos
        - company
      properties:
        jobNumber:
          type: string
        customerName:
          type: string
        address:
          type: string
        city:
          type: string
        flooringType:
          $ref: "#/components/schemas/FlooringType"
        crewAssigned:
          $ref: "#/components/schemas/CrewAssigned"
        squareFootage:
          type: number
        estStartDate:
          type: string
        estCompletionDate:
          type: string
        status:
          $ref: "#/components/schemas/JobStatus"
        stageHistory:
          type: array
          items:
            $ref: "#/components/schemas/StageEvent"
        photos:
          type: array
          items:
            $ref: "#/components/schemas/JobPhoto"
        quote:
          $ref: "#/components/schemas/PublicPortalQuote"
          nullable: true
        invoices:
          type: array
          items:
            $ref: "#/components/schemas/PublicInvoice"
        balance:
          $ref: "#/components/schemas/PublicBalance"
        company:
          $ref: "#/components/schemas/PublicCompany"
    PublicInvoice:
      type: object
      required:
        - id
        - invoiceNumber
        - lineItems
        - subtotal
        - taxableAmount
        - taxAmount
        - discountAmount
        - total
        - depositAmount
        - paidAmount
        - balanceAmount
        - refundedAmount
        - taxCode
        - paymentReference
        - status
        - issueDate
        - dueDate
        - notes
      properties:
        id:
          type: string
        invoiceNumber:
          type: string
        lineItems:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceLineItem"
        subtotal:
          type: number
        taxableAmount:
          type: number
        taxAmount:
          type: number
        discountAmount:
          type: number
        total:
          type: number
        depositAmount:
          type: number
        paidAmount:
          type: number
        balanceAmount:
          type: number
        refundedAmount:
          type: number
        taxCode:
          type: string
        paymentReference:
          type: string
        paidAt:
          type: string
        status:
          $ref: "#/components/schemas/InvoiceStatus"
        issueDate:
          type: string
        dueDate:
          type: string
        notes:
          type: string
    PublicBalance:
      type: object
      required:
        - invoicedTotal
        - paidTotal
        - outstanding
      properties:
        invoicedTotal:
          type: number
        paidTotal:
          type: number
        outstanding:
          type: number
    PublicPortalQuote:
      type: object
      required:
        - shareToken
        - customerName
        - projectLocation
        - flooringType
        - roomList
        - totalSqFt
        - scopeOfWork
        - estimatedPrice
        - expectedTimeline
        - materialAssumptions
        - exclusions
        - warrantyNote
        - depositType
        - depositValue
        - depositAmount
        - paymentTerms
        - status
        - signature
      properties:
        shareToken:
          type: string
        customerName:
          type: string
        projectLocation:
          type: string
        flooringType:
          $ref: "#/components/schemas/FlooringType"
        roomList:
          type: array
          items:
            $ref: "#/components/schemas/Room"
        totalSqFt:
          type: number
        scopeOfWork:
          type: string
        estimatedPrice:
          type: number
        expectedTimeline:
          type: string
        materialAssumptions:
          type: string
        exclusions:
          type: string
        warrantyNote:
          type: string
        depositType:
          $ref: "#/components/schemas/DepositType"
        depositValue:
          type: number
        depositAmount:
          type: number
        paymentTerms:
          type: string
        status:
          $ref: "#/components/schemas/ProposalStatus"
        acceptedAt:
          type: string
        declinedAt:
          type: string
        signature:
          type: string
...
    PublicQuote:
      type: object
      required:
        - customerName
        - projectLocation
        - flooringType
        - roomList
        - totalSqFt
        - scopeOfWork
        - estimatedPrice
        - expectedTimeline
        - materialAssumptions
        - exclusions
        - warrantyNote
        - depositType
        - depositValue
        - depositAmount
        - paymentTerms
        - status
        - signature
        - company
      properties:
        customerName:
          type: string
        projectLocation:
          type: string
        flooringType:
          $ref: "#/components/schemas/FlooringType"
        roomList:
          type: array
          items:
            $ref: "#/components/schemas/Room"
        totalSqFt:
          type: number
        scopeOfWork:
          type: string
        estimatedPrice:
          type: number
        expectedTimeline:
          type: string
        materialAssumptions:
          type: string
        exclusions:
          type: string
        warrantyNote:
          type: string
        depositType:
          $ref: "#/components/schemas/DepositType"
        depositValue:
          type: number
        depositAmount:
          type: number
        paymentTerms:
          type: string
        status:
          $ref: "#/components/schemas/ProposalStatus"
        acceptedAt:
          type: string
        declinedAt:
          type: string
        signature:
          type: string
        company:
          $ref: "#/components/schemas/PublicCompany"
    QuoteAcceptInput:
      type: object
      required: [signature]
      properties:
        signature:
          type: string
          minLength: 1
...
    InvoiceLineItem:
      type: object
      required: [id, description, category, quantity, unitPrice]
      properties:
        id:
          type: string
        description:
          type: string
        category:
          $ref: "#/components/schemas/InvoiceLineCategory"
        quantity:
          type: number
        unitPrice:
          type: number
...